+++ packages/harmony_log/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.1
+
+- Added `dart:developer` plain logger using `LogPlainOutput.developerLog()`.
+
 ## 2.0.0
 
 - **BREAKING**: Using `child` in `Log` instead of `output`.
@@ -6,7 +10,7 @@
 - Removed `uuid` log id generation.
 - Added noop log implementation `Log.noop()`.
 - Added `Zone` information to log events.
-- Added `dart:developer` logger using `LogOutput.developerLog`.
+- Added `dart:developer` logger using `LogOutput.developerLog()`.
 
 ## 1.0.0